frontend: Add explorer URL to wallet
When adding a chain to the connected wallet, the explorer URL is now added. This also changes our config files by renaming `explorerTransactionUrl` to `explorerUrl` and removing the `/tx` path from the urls. This can be safely done as explorer routes are standardized in EIP-3091. Furthermore, the `addChain` was changed in that way that it now consumes the usual `Chain` type instead of the `ChainData` type. The `ChainData` type was added when we did not have the split types of the configuration and is redundant now. Same goes for `Token` and `TokenData`.