frontend: refactor & tests for fill-manager service
The fill-manager service contained code that can be considered
a cross-app utility.
Mainly the `getContract` and `getCurrentBlockNumber`
functions should be implemented in a general/reusable manner
for the whole code base and not fill-manager specific.
Same goes for the historic event filtering until first match.
A function like this can be used by any other service to try to find a
match in the history of events based on a provided filter.
This commit provides tests for the newly created files, fill-manager
service & adjusts the tests for the other files that were affected
by the refactoring.