frontend: correct chain files
Metamask compares the chain we want to add to the records on this site: https://chainlist.wtf . if the name, rpc or any other info is different than what is returned there it displays a warning to the user. By having the same values as returned by the site the user is not confused by metamask warnings anymore.

frontend: add nativeCurrency to all configs