frontend: extend transaction error message handler
Previously, whenever a `createRequest` or `withdraw` transaction would
fail, the message displayed to the user was a static
string "Insufficient balance."
This already caused a lot of confusion. A temporary solution
(until an error handling module is implemented) would be to
return the reason returned from the contract assertions (if available).