frontend: Hotfix for MetaMask throwing disconnect events when switching chain
MetaMask sometimes throws a disconnect event when switching the chain. In this case we need to ignore the event, because the account is actually still connected. This adds a flag that is true for 1 second after a chain switch happened. A short time period is necessary because the event is not directly thrown when the `wallet_switchEthereumChain` rpc method returns. If this is fixed on the side of MetaMask, we can remove the flag again.