frontend: Fix fee query on transfer creation
When creating a transfer, the request manager was queried for the fees of the target token address (which only exists on the target chain). This obviously returns an erroneous value as the request manager only knows about the source token address. This led to the bug that the fees were not taken into account for the getting the token allowance. This simply changes the `getRequestFee` call to use the correct variable.