frontend: improve ui
The font-size on the page was set to 50% which was causing some buttons to be barelly readable on screens with high resolution. In adition to that one had to zoom the page to get it to a comfortable size.

I’ve moved the whole container lower on the page -> it is now bellow the "logo area". The logo container is now also restricted in width so that the logo won’t appear on the far left of the page and the content of the page would be in the middle.

The page is now also more responsive - on a bigger browser the actual content is wider.