frontend: implement real-time claim count watcher
The request life-cycle changed in the contracts v1.
In the previous version of the contracts, requests that have expired
couldn't be claimed anymore.
In the new version requests can be claimed, even after expiry,
until a specific moment in the future.
This means that we need to watch and react on the `ClaimMade` &
`ClaimStakeWithdrawn` events in order to update the transfer state
so we can disable withdrawing until the claims have been withdrawn.
This mostly focuses on the case where a request has expired without
a fill event but yet someone claims to have filled the event.
The withdraw button will only be visible when all the claims have been
withdrawn / dispute has been resolved.