relayer: reduce max message length
Based on a prior message byte size research we decided to
lower down the required deposit (in the relayer) on ArbitrumL1Messenger
contract for the process of relaying a message to Arbitrum.
Previously, the idea was, to not depend on a prior knowledge of
the possible sizes of our messages (which can be determined if one
looks at the Resolver contract)
and just use the max allowed message length on Ethereum (1 MB).
But this resulted in extremely high deposit amounts and we believe that
is not necessary. So we had to switch to the other option.
The value (5000 bytes) that this commit introduces is basically
derived by multiplying the message size, which is being sent via the
Resolver, by 10.