frontend: extend allowed node engine versions
Not a long time ago, we decided to move all our modules to Node 18
due to it being the current LTS. Node 16 went into maintenance and
soon it will enter a non-maintained phase.
This move came with 1 unforeseen drawback:
For hosting our frontend we are currently using DO App.
DO App's don't support Node 18 so we have to allow Node 16 as supported
version for the time being.
Our app works perfectly fine with Node 16 so no need to impose hard
restrictions here.