frontend: optimize request transfer
Previously, the code didn't account for the fact that it takes a certain
amount of time from the moment
a user initiates a transfer request until he actually submits it via
his wallet provider.
The transfer object was fetching & storing the target chain block at the
moment of triggering a "createRequest" transaction call
instead of after successful submission of the transaction call.