frontend: deal better with reverted txs
We were fetching the transaction receipt and were always assuming that if we have a tx receipt we have a succesfull tx. In some situations though, we had reverted transaction. We were trying to parse the logs on those, which triggered a js error: `Cannot read properties of undefined (reading 'topics')`

We now check the receipt status and try to parse the logs only if it is succesful.

frontend: Deal better with reverted TXs