frontend: prevent user from submitting a second tx
If the user has another transfer pending, we prevent the user from submitting a second transfer.

We don’t treat a transfer to be pending if the user has signed all transactions and they’ve been mined. If that is the case and an agent still hasn’t fullfiled the request - this is not a pending transaction.

The reason we do that is that before submiting a transaction we check the allowance and create an allowance tx. But if the actual transaction that submit the money to the contract is not mined yet, then the second tx that the user starts has an incorrect allowance and the tx that sends the money to the contract fails with "insuficient allowance".