frontend: fix misconfigured code coverage
After a recent update of the vitest library from 0.9.3 to 0.12.10 the
code coverage was broken. The `yarn test:coverage` was not
outputting anything.
The reason for this was a change in the config interface that is used
inside vite.
Also, the yarn.lock file had 2 different versions of vite installations
so the interfaces of both version were conflicting.