frontend: fix provider imports
All the wallet provider modules were importing their dependencies from
the index file. Since prettier is enforcing its sorting strategy on
the imports we lost the correct order of imports in the modules.
In order to avoid this conflict we are now importing the types & other
deps by specifying the precise paths to the files.