frontend: fix metamask <-> coinbase conflict
There is the case where users might have multiple wallet provider
extensions installed in their browsers.
In such cases, the `window.ethereum` object is being transformed
into an object that holds all these providers instead of
representing the single provider.
By default the selected provider is the first provider found in the
array `window.ethereum.providers`. However, we want to be specific
about which extensions we use when creating our internal provider
instances.
This commit solves this conflict and makes sure we are instantiating
the `EthereumProvider` with the correct state.