frontend: Fix wallet disconnection after chain switch
Slow RPCs can still lead to the problem that after a chain switch a wallet disconnection occurs when using MetaMask. Looks like the timeout of 1 second is not enough. To fix it deterministically, I removed the listener on the disconnect event completely for the MetaMaskProvider. The listener was used to catch the case when a user clicks disconnect inside MetaMask. In this case also an accountsChanged event is fired. We can simply rely on this event to call our disconnect function. I also removed the no disconnection after chain switch test as it is not applicable anymore.