beamer: health: add health-check script
The health-check script scans the contracts for the emitted events and analyzes whether there is a missed fill, unclaimed transaction or a challenge game going on.

In addition to this if an agent-address is provided in the conf it will fetch that agent’s balance from the chain.

The script doesn’t output to stdout, but sends everything either to telegram or rocketchat.