frontend: Ensure chainId is changed after chain switch
Checks the chainId of the provider and copies its value after a chain switch triggered by the dapp. Also moves the network switch listening to our EthereumProvider, because I think this logic has no relation to the RequestDialog where it was before. Also now uses the network event of ethers.js. This was more reliable than the one of the window.ethereum API.