contracts: add automatic withdrawal in messenger
Here we want to optimize on gas costs and agent experience.
Previously, the agents had to keep track how much funds
they have deposited in the ArbitrumMessenger contract
and respond accordingly(withdraw) when they want to stop providing
liquidity for this chain.
With the automatic withdrawal they wont have to worry about
this.
Also, from the fee perspective, we are reducing the number of
transactions they need to do by 1, which leads to less fees spent.