Add a .nvmrc file
A lot of people use NVM to manage node environments. We are generally enforcing node engine 18.x in the main package.json and in the relay’s package.json. For the frontend we say node > 16.

NVM has the option to specify a node version to use in an .nvmrc file. By having this file one can now just do:
```
nvm use
```

Or even have a deeper shell integration and the moment one enters the directory nvm could load the specified node version. For more info on how to set this up:
https://github.com/nvm-sh/nvm#nvmrc