frontend: extend transfer objects with timestamps
As we are working on new features for the app
where the timestamp of a transfer creation
& transfer fill events are required
there was a need of extending the current Transfer class.
Extended its implementation with a new field to encapsulate
the state for the fill event, while reused the current structure
to store the request creation timestamp.
The `fetchUntilFirstMatchingEvent` was only returning a boolean value
indicating whether the event was found or not but didn't return any
valuable information that would enable us to use that specific event
object we are querying for.
I improved this by adding proper types and adapting the function to
return an event instead of a boolean.