frontend: Adapt configure script to new deployments package
The deplyoments package uses a new structure for deployment json files. This also removes all unused properties of the deployment file type declarations. I did this because the structure of the deployment JSON files saw frequent changes in the past. Declaring only the used properties makes it easier to adapt to future changes of the deployments structure.