deployments: fix missing npm package
I initially had thought of testing the package right after it is being created, but this idea changed. We now have a test step, that generates the packages, tests it and removes it.

The test also works on the dist folder and now since the test runs after the package creation it removes the package which causes the next step in the circleCI flow to fail as it cannot find the package file for publishing.