frontend: improve continuation of interrupted transfers
Previously, the app had no notion of pending transactions/signature
prompts.
When certain actions of the transfer objects were in pending
state (waiting for user signature) and in the meantime the user's
browser sesssion was terminated, the app was not able to pick up
from the last completed action properly.
Each time the user submitted his signature and closed the browser tab,
the app kept no history of the transaction hashes of all the actions.
This means that when he comes back on our app, he will be prompted to
submit a transaction for the pending actions of a transfer.
This is now fixed and users can freely close and open the app without
worrying that the app will lose the transaction history related to
a transfer.