frontend: Use new deployments package for contract ABIs and types
With the deployments package we now have separate ABIs for our mainnet and goerli deployments. Before we just had one version of the ABIs which were just copied to the frontend folder. So the correct ABIs must get loaded depending on the mode in which Vite is running. Mainnet for production, Goerli for staging / development. First, I tried dynamic imports for achieving this. These were adding asynchronity to functions which were not async before. I didn't like that and it would have needed refactoring. I found a pretty new way to achieve environment-specific imports here: https://github.com/vitejs/vite/discussions/4172#discussioncomment-5496820 . It uses the `imports` field of package.json to define which folder to use in the different modes. It makes sure that only one version of the ABIs goes into the final bundle. It required to update Typescript to v5.

For the token, we only have the ABI in the goerli deployment folder. So we always use this. Should be ok, as it should never change.