e2e: Retry relayer if exited with an error
There might be the case in the e2e tests that a message was not finalized yet and the relayer is already started. To align with the agent, the e2e tests should restart the relayer. For the implementation, I separated the execution of the relayer in a new common shell function.