relayer: add message proving functionality
The relayer component had to be extended with a separate
sub-command for proving OP messages on L1.
This process is separate from the process of relaying and therefore
moved to its own subcommand.
The proving functionality/responsibility is still
a part of the OptimismRelayerService
and therefore implemented as a part of it even though it is used
in different processes.
This will enable for its reusability.