relayer: fix broken pkg builds
Since we introduced multiple subcommands and
handed over the executable path parsing responsibility to the
commander.js library, we lost support for the binary package
bundling of the relayer due to
failed module resolution.
A slight refactor was necessary in order to get this setup working.
While at it, i saw the opportunity to even further refactor, clean up
code repetition and improve the design of our CLI.
This new CLI design also enables deep "--help" documentation for each
sub-command.