relayer: use static 3M gasLimit for OP relays
We noticed that the gas estimation with the usage of the
"hack" (introduced in commit `23fe2f2`) will not work in some cases.
For example: for relaying a message from OP to ARB we
assert that the `tx.origin` has enough funds deposited in the
ArbitrumL1Messenger for relaying the message successfully.
However, since in the "hack" we use `0x0..01` address as `from` address
the estimation will fail and revert at the above mentioned assertion.
We decided to go back to setting a fixed gasLimit value that is high
enough so we avoid calling the `estimateGas` functions for the
calls on the OP contracts.