relayer: fix flaky e2e test
There was a bug in nitro where a batch wouldn’t be posted until an additional transaction was submitted to the L2 before the max batch posting delay took effect. One has to wait for the max posting batch delay and then submit a new transaction. This bug is fixed in 2.1.0-beta.4 so updating to this version fixes the issue.

More info here https://github.com/OffchainLabs/nitro/issues/1706

Despite us setting the checkout version of the submodule to the 2.1.0-beta.4 tag one needs to update the NITRO_NODE_VERSION in the test-node.bash script as there it is loading 2.1.0-beta.1