frontend: properly configure WC 2
The previous configuration was providing all the chains as required
chains. This means that when the user wants to connect to our
app via WC 2, he needs to have all the chains that we
support (at that moment) configured in his wallet app.
WC 2 provides a way to specify required/optional chains and here we
are using that.
By specifying most of the chains as optional, the user only needs to
have one main chain configured (Ethereum)
and the rest will be taken care of by our app
(adding chains/switching chains).