frontend: adjust `generate-types` command
Previously, we only needed to generate typings for the
@beamer-bridge/deployments npm package. That changed with the new
`FeeSub` contract. The decision was to not keep this contract as a part
of our deployments folder/npm package so
we have to copy/paste the ABI for this contract
whenever it changes.
Due to this reason, we had to extend the `generate-types`
npm command to detect and use this folder as well.
In order to keep the output folder structure of this command simple,
i used the `typechain` command twice so it doesn't try to come up with
a complex folder structure for the outputted contracts which breaks
the module import resolution in our codebase.