frontend: `SubsidizedTransfer`: Only overwrite transaction wrapper calls
In the former implementation things could easily go wrong because the functions of the `Transfer` class are quite big. It was not easily seeable what happens when the `requestManagerAddress` was overwritten by the `feeSubAddress`. The new implementation should be fail-safer. Only the necessary transaction wrappers are overwritten by providing the `feeSubAddress` in the call to the transaction function.