beamer: agent: take the fill mutex before doing anything else when filling
Otherwise it can happen that we do the timing checks before we obtain the mutex,
and we are forced to wait on the mutex long enough that the timing check
performed previously becomes invalid. Same reasoning applies to other
checks, e.g. the min_source_balance, token allowance check etc.