deployments: create package with abi bytecode
Breaking change!

Currently we generate the abis without the bytecode. We do this because the frontend doesn’t use the bytecode. However the agent and backend code needs the bytecode.

The abis now generated in /abis/ will have the bytecode by default.
If someone needs abis without the bytecode they are in the abis-without-bytecode folder.