beamer: config: Ensure correct token keys
Either token key is a symbol and exists in token_addresses or it is a checksum address.
Also check that token_addresses stays unmodified as it is autogenerated from on chain data.
Additionally, if checksum address check that address is not in token_addresses.
Only add to token_addresses if it is actually a symbol to be resolved to the address.